+++ .gitignore
@@ -1,0 +1,9 @@
+*.dll
+*.pdb
+Decisions.GoogleDrive.TestSuite/bin/
+Decisions.GoogleDrive.TestSuite/obj/
+Decisions.GoogleDrive/bin/
+Decisions.GoogleDrive/obj/
+Decisions.GoogleDrive/packages/
+Decisions.GoogleDrive/.vs
+
