+++ Decisions.GoogleDrive/Decisions.GoogleDrive.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Decisions.GoogleDrive", "Decisions.GoogleDrive.csproj", "{C3CEB50E-4797-4A22-86A5-8060E610447C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Decisions.GoogleDrive.TestSuite", "..\Decisions.GoogleDrive.TestSuite\Decisions.GoogleDrive.TestSuite.csproj", "{E4A927B7-ACF1-42E9-B5DD-B7BEC7C3D8E6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C3CEB50E-4797-4A22-86A5-8060E610447C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C3CEB50E-4797-4A22-86A5-8060E610447C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C3CEB50E-4797-4A22-86A5-8060E610447C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C3CEB50E-4797-4A22-86A5-8060E610447C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E4A927B7-ACF1-42E9-B5DD-B7BEC7C3D8E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E4A927B7-ACF1-42E9-B5DD-B7BEC7C3D8E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E4A927B7-ACF1-42E9-B5DD-B7BEC7C3D8E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E4A927B7-ACF1-42E9-B5DD-B7BEC7C3D8E6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7C2427D3-B85D-4AE5-856F-766AB22CB327}
+	EndGlobalSection
+EndGlobal
