--- DriveLibraryTests/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Google.Apis" version="1.45.0" targetFramework="net48" />
-  <package id="Google.Apis.Auth" version="1.45.0" targetFramework="net48" />
-  <package id="Google.Apis.Core" version="1.45.0" targetFramework="net48" />
-  <package id="Google.Apis.Drive.v3" version="1.45.0.1922" targetFramework="net48" />
-  <package id="MSTest.TestAdapter" version="2.1.0" targetFramework="net48" />
-  <package id="MSTest.TestFramework" version="2.1.0" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net48" />
-</packages>
