+++ Module.Build.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0"?>
+<CreateModuleInfo xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <Clients>
+    <DotNetCoreDll>.\packages\DecisionsSDK.6.8.0\lib\CoreServicesClients.Net.dll</DotNetCoreDll>
+    <DotNetFrontendFrameworkDll>.\packages\DecisionsSDK.6.8.0\lib\DecisionsFramework.Net.dll</DotNetFrontendFrameworkDll>
+    <SLCoreDll></SLCoreDll>
+    <SLFrontendFrameworkDll></SLFrontendFrameworkDll>
+    <SilverLightsSDKAssembliesPath></SilverLightsSDKAssembliesPath>
+    <References/>
+  </Clients>
+  <Resources NoComponents="true">
+    <ServicesDlls>
+      <string>.\Decisions.GoogleDrive\bin\Debug\Decisions.GoogleDrive.dll</string>
+    </ServicesDlls>
+    <CliDlls/>
+    <WebPages/>
+	<MvcViews>
+	</MvcViews>
+	<MvcDlls>
+	</MvcDlls>
+    <AgentDlls>
+    </AgentDlls>
+    <InstallScript>
+    </InstallScript>
+    <SqlScripts />
+    <ObjectsToImport></ObjectsToImport>
+    <MsSqlScript>
+    </MsSqlScript>
+    <MySqlScript>
+    </MySqlScript>
+    <OracleScript>
+    </OracleScript>
+  </Resources>
+  <Description>Module to manage files and folders in GoogleDrive.</Description>
+</CreateModuleInfo>
