+++ build.proj
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project
+    xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0" DefaultTargets="build_all">
+    <PropertyGroup>
+        <CreateModule>"C:\Program Files\Decisions\Decisions Services Manager\CreateModule.exe"</CreateModule>
+    </PropertyGroup>
+    <Target Name="build_all">
+        <MSBuild Projects=".\Decisions.GoogleDrive\Decisions.GoogleDrive.csproj" />
+        <CallTarget Targets="build_module"/>
+    </Target>
+    <Target Name="build_module">
+        <Exec Command="$(CreateModule) -buildmodule Decisions.GoogleDrive -output &quot;.&quot; -buildfile Module.Modified.xml"/>
+    </Target>
+</Project>
